[ADF-123] [BE] Cập nhập logic DBContext trong Repository

+++ Repositories/Implement/PackageRepository.cs
@@ -3,37 +3,55 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Repositories
 {
-	public class PackageRepository : SingletonBase<PackageRepository>, IPackageRepository
-	{
-		public async Task CreateList(List<Package> package)
-		{
-				await context.Packages.AddRangeAsync(package);
-				await context.SaveChangesAsync();	
-		}
-
-		public async Task Delete(Guid id)
-		{
-			var package = await context.Packages.FirstOrDefaultAsync(i => i.Id == id);
-			context.Packages.Remove(package);
-			await context.SaveChangesAsync();
-		}
-
-		public async Task<IEnumerable<Package>> GetAlls()
-		{
-			var packages = await context.Packages.ToListAsync();
-			return packages;
-		}
-
-		public async Task<Package> GetById(Guid id)
-		{
-            var package = await context.Packages.FirstOrDefaultAsync(i => i.Id == id);
-            return package!;
+    public class PackageRepository : IPackageRepository
+    {
+        public async Task CreateList(List<Package> packages)
+        {
+            using (var context = new PostgresContext())
+            {
+                await context.Packages.AddRangeAsync(packages);
+                await context.SaveChangesAsync();
+            }
         }
 
-		public async Task Update(Package package)
-		{
-			context.Entry(package).State = EntityState.Modified;
-			await context.SaveChangesAsync();
-		}
-	}
+        public async Task Delete(Guid id)
+        {
+            using (var context = new PostgresContext())
+            {
+                var package = await context.Packages.FirstOrDefaultAsync(i => i.Id == id);
+                if (package != null)
+                {
+                    context.Packages.Remove(package);
+                    await context.SaveChangesAsync();
+                }
+            }
+        }
+
+        public async Task<IEnumerable<Package>> GetAlls()
+        {
+            using (var context = new PostgresContext())
+            {
+                var packages = await context.Packages.ToListAsync();
+                return packages;
+            }
+        }
+
+        public async Task<Package> GetById(Guid id)
+        {
+            using (var context = new PostgresContext())
+            {
+                var package = await context.Packages.FirstOrDefaultAsync(i => i.Id == id);
+                return package!;
+            }
+        }
+
+        public async Task Update(Package package)
+        {
+            using (var context = new PostgresContext())
+            {
+                context.Entry(package).State = EntityState.Modified;
+                await context.SaveChangesAsync();
+            }
+        }
+    }
 }
